admin can delete news
