+++ frontEnd/src/App.jsx
@@ -8,8 +8,8 @@ import Teams from "./pages/teams.jsx";
 import ShopPage from './pages/ShopPage.jsx';
 import NewsPage from './pages/news.jsx';
 import SingleNews from "./components/singleNews.jsx";
-import AddNews from "./components/admin/addNews.jsx";
 import EditNews from "./components/admin/editNews.jsx";
+import EditSingleNews from "./components/admin/editSingleNews.jsx";
 import EditPlayerList from "./components/admin/editPlayerList.jsx";
 import EditShopItem from "./components/admin/editShopItem.jsx";
 import Login from "./pages/loginPage.jsx";
@@ -26,8 +26,8 @@ const App = () => {
       <div style={{ marginTop: "100px" }}>
         <Routes>
           <Route path="/dashboard" element={user ? <Dashboard /> : <Navigate to={"/login"}/>} />
-          <Route path="/adminDashboard/addNews" element={user ? <AddNews /> : <Navigate to={"/login"}/>} />
-          <Route path="/adminDashboard/addNews/:id" element={user ? <EditNews /> : <Navigate to={"/login"}/>} />
+          <Route path="/adminDashboard/editNews" element={user ? <EditNews /> : <Navigate to={"/login"}/>} />
+          <Route path="/adminDashboard/editNews/:id" element={user ? <EditSingleNews /> : <Navigate to={"/login"}/>} />
           <Route path="/adminDashboard/editShop" element={user ? <EditShopItem /> : <Navigate to={"/login"}/>} />
           <Route path="/adminDashboard/editPlayers" element={user ? <EditPlayerList /> : <Navigate to={"/login"}/>} />
           <Route path="/adminDashboard" element={user ? <AdminDashboard /> : <Navigate to={"/login"}/>} />
