--- frontEnd/src/allCss/addNews.css
@@ -1,5 +1,0 @@
-.divider {
-    height: 10px;
-    background-color: black;
-    margin: 20px 0px 20px 0px; 
-}
